move guarantee code to separate file
